refactor(herbatika): extract product gallery state
